@@ -1,12 +1,9 @@
 name: code_size_package
-description: A new Flutter project.
-
-publish_to: "none"
-
-version: 1.0.0+1
+description: A Flutter project demonstrating code size issues with packages.
+publish_to: none
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ^3.0.0
 
 dependencies:
   flutter:
